# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/gamification_goal_definitions:
    get:
      tags:
        - GamificationGoalDefinitions
      summary: List GamificationGoalDefinitions
      description: Get a paginated list of GamificationGoalDefinitions with optional filtering and sorting
      operationId: listGamificationGoalDefinitions
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GamificationGoalDefinitionList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - GamificationGoalDefinitions
      summary: Create GamificationGoalDefinition
      description: Create a new GamificationGoalDefinition
      operationId: createGamificationGoalDefinition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGamificationGoalDefinitionRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GamificationGoalDefinition'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    GamificationGoalDefinition:
      type: object
      required:
        - id
        - name
        - monetary
        - computation_mode
        - condition
        - display_mode
        - metadata
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 200
        description:
          nullable: true
          type: string
        suffix:
          nullable: true
          type: string
          maxLength: 40
        monetary:
          type: boolean
        computation_mode:
          type: string
        metric_key:
          nullable: true
          type: string
          maxLength: 120
        condition:
          type: string
        display_mode:
          type: string
        metadata:
          type: object
...
    CreateGamificationGoalDefinitionRequest:
      type: object
      required:
        - name
        - monetary
        - computation_mode
        - condition
        - display_mode
        - metadata
      properties:
        name:
          type: string
          maxLength: 200
        description:
          nullable: true
          type: string
        suffix:
          nullable: true
          type: string
          maxLength: 40
        monetary:
          type: boolean
        computation_mode:
          type: string
        metric_key:
          nullable: true
          type: string
          maxLength: 120
        condition:
          type: string
        display_mode:
          type: string
        metadata:
          type: object
...
    GamificationGoalDefinitionList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/GamificationGoalDefinition'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []